requirements.txt
anthropic>=0.42.0
edge-tts
ffmpeg-python
pillow
python-dotenv
psutil
